feat: Integrando o front-end ao back-end

+++ backend/src/main/java/com/clinicavitalis/backend/patient/PatientRequestDTO.java
@@ -4,8 +4,8 @@ public record PatientRequestDTO(
         String name, 
         String cpf, 
         String birthdate,
-        Long weight, 
-        Long height, 
+        Double weight, 
+        Double height, 
         String uf
     ) {
